[Fix] delete pollution, code & layout fixes (#10)
* fix: remove excess brackets

* refactor: remove useless function `research()`

`research()` is a useless and confusing function.

Firstly, its job is to call `getWeather()` when it's called, so that replacing it with `getWeather()` is not going to affect anything.

Secondly, its work doesn't match its name, "research"; it actually does "search" job, which creates confusion.

* fix: fix the layout, after adding coords

* fix: temporarily disable pollution

We are not prepared enough to add pollution to our app :(

* fix: bring the 404 error page back :)

* fix: remove default value for coords